move Canvas to new file
